avoid cast of float to double in SplitterQuadratic

+++ src/main/java/com/github/davidmoten/rtree/SplitterQuadratic.java
@@ -73,9 +73,9 @@ public final class SplitterQuadratic implements Splitter {
     static <T extends HasGeometry> T getBestCandidateForGroup(List<T> list, List<T> group,
             Rectangle groupMbr) {
         Optional<T> minEntry = absent();
-        Optional<Double> minArea = absent();
+        Optional<Float> minArea = absent();
         for (final T entry : list) {
-            final double area = groupMbr.add(entry.geometry().mbr()).area();
+            final float area = groupMbr.add(entry.geometry().mbr()).area();
             if (!minArea.isPresent() || area < minArea.get()) {
                 minArea = of(area);
                 minEntry = of(entry);
@@ -89,12 +89,12 @@ public final class SplitterQuadratic implements Splitter {
         Optional<T> e1 = absent();
         Optional<T> e2 = absent();
         {
-            Optional<Double> maxArea = absent();
+            Optional<Float> maxArea = absent();
             for (int i = 0; i < items.size(); i++) {
                 for (int j = i + 1; j < items.size(); j++) {
                     T entry1 = items.get(i);
                     T entry2 = items.get(j);
-                    final double area = entry1.geometry().mbr().add(entry2.geometry().mbr()).area();
+                    final float area = entry1.geometry().mbr().add(entry2.geometry().mbr()).area();
                     if (!maxArea.isPresent() || area > maxArea.get()) {
                         e1 = of(entry1);
                         e2 = of(entry2);
